requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.9.0

# Machine learning
optuna>=3.3.0
umap-learn>=0.5.4
imbalanced-learn>=0.11.0

# Chemistry and molecular
rdkit>=2023.3.1
gensim>=4.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Deep learning
tensorflow>=2.13.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
pyyaml>=6.0
loguru>=0.7.0
click>=8.1.0

# Web app
streamlit>=1.28.0

# Development
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0